Refactored DBs and Created Sales view
Refactored DBs to store all the information a user might want to know about each item,  with enough additional information to be able to analyze it later.

Created ItemTable to allow for easier table generation on the UI.

Renamed Index to manage for clarity with design.

Refactored some HTML files so they look better or have more clear instructions.