Began work on post-items
Realized the post date is initialized in the test data, changed that so items in item_data are initialized without a post_date.

Began reading in item information from item_data.db
If the post_date is null it means it is not posted, thus iterate through those items and begin the posting process. Marked where code is believed to stop working.