Worked on post-item
Got post-item working up until it has to set the size.

The options for the item details change depending on what the item actually is, ie shoes have different fields from jewelry.

That means each category of item will need specialized info that other categories may not share. This could be tough to manage, so it will require some thought.