Completed pull_selects.py
The pull_selects.py can now be run to automatically pull down the categories and subcategories from the grailed site, saving them under 'config/selects.json' for later use.

These selects can now be used to populate the post.html page with the options.

Sizing and other considerations that are specific to each category/subcategory still need to be pulled down and populated.