Finished designers and categories for pull_selects
pull_selects.py now pulls down the list of designers and categories that can be sold on grailed. designers are saved in the file 'scripts/configs/designers'

There is still the need for unique sizing information depending on the category, but this should work for the refactoring of post.html w.r.t. category and designer.

I also introduced a new text document that briefly (and poorly) describes some troubleshooting tips I have run into while working on the project.

In the process of writing this update my chromedriver also needed to be updated (which also spurred on the troubleshooting)