Refactored and improved
Refactored the post_item route to get rid of bug where the items would not be uploaded to the database.

Replaced the use of Item tables with a much more straightforward approach to querying the dbs and getting the data as json. The json is then converted to a table. This eliminated the need for ItemTable.py

The code that makes the call to the database and converts the returned data is used twice in flask_site.py and can be refactored into a method for repeated invokation.

Changed the javascript to hold an empty value for the input box so it won't store a string as the value when left empty.